Transfer params checking in ConsoleClient
