+++ TagCloudContainer/Program.cs
@@ -47,8 +47,8 @@ namespace TagCloudContainer
                             (pi, ctx) => pi.Name == "backgroundColor",
                             (pi, ctx) => ctx.Resolve<IUserInterface>().Config.BackgroundColor),
                         new ResolvedParameter(
-                            (pi, ctx) => pi.Name == "imageExtension",
-                            (pi, ctx) => ctx.Resolve<IUserInterface>().Config.ImageExtension)
+                            (pi, ctx) => pi.Name == "imageFormat",
+                            (pi, ctx) => ctx.Resolve<IUserInterface>().Config.ImageFormat)
                     })
                 .As<ITagCloudVisualization>();
